shachotwo-app/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
supabase>=2.0.0
google-genai>=1.0.0
anthropic>=0.40.0
openai>=1.50.0
voyageai>=0.3.0
python-multipart>=0.0.9
redis>=5.0.0
pypdf>=4.0.0
openpyxl>=3.1.0
python-docx>=1.1.0
python-pptx>=1.0.0
gunicorn>=22.0.0
python-dotenv>=1.0.0
cryptography>=42.0.0
pytest-asyncio>=0.24.0
beautifulsoup4>=4.12.0
json5>=0.9.0
json-repair>=0.25.0
langgraph>=0.2.0
